Even more progress

+++ generic/itclClass.c
@@ -1407,7 +1407,7 @@ int
 Itcl_HandleClass(
     void *clientData,        /* class definition */
     Tcl_Interp *interp,      /* current interpreter */
-    int objc,                /* number of arguments */
+    size_t objc,                /* number of arguments */
     Tcl_Obj *const objv[])   /* argument objects */
 {
     if (objc > 3) {
